+++ src/Core/Operations/Updater.php
@@ -4,6 +4,7 @@ namespace WP_Statistics\Core\Operations;
 
 use WP_Statistics\Core\AbstractCore;
 use WP_Statistics\Components\AssetNameObfuscator;
+use WP_Statistics\Components\DateTime;
 use WP_Statistics\Components\Event;
 use WP_Statistics\Components\SystemCleaner;
 use WP_STATISTICS\DB;
@@ -325,11 +326,19 @@ class Updater extends AbstractCore
          * Update consent integration to WP Consent API for backward compatibility
          */
         $integration          = Option::get('consent_integration');
-        $consentLevel         = Option::get('consent_level_integration', 'disabled');
+        $consentLevel         = Option::get('consent_level_integration', 'functional');
         $isWpConsentApiActive = IntegrationHelper::getIntegration('wp_consent_api')->isActive();
 
         if ($isWpConsentApiActive && empty($integration) && $consentLevel !== 'disabled') {
             Option::update('consent_integration', 'wp_consent_api');
+        }
+
+        /**
+         * Unset consent integration if consent level is disabled
+         */
+        if ($integration === 'wp_consent_api' && $consentLevel === 'disabled') {
+            Option::update('consent_integration', '');
+            Option::update('consent_level_integration', 'functional');
         }
 
         /**
@@ -388,6 +397,14 @@ class Updater extends AbstractCore
             if (!empty($updatedExcludedUrls)) {
                 Option::update('excluded_urls', implode("\n", $updatedExcludedUrls));
             }
+        }
+
+        if (!Option::get('schedule_dbmaint') && version_compare($this->currentVersion, '14.15.5', '<')) {
+            Option::update('schedule_dbmaint_days', '0');
+        }
+
+        if (Option::get('schedule_dbmaint') && version_compare($this->currentVersion, '14.15.5', '<')) {
+            Event::reschedule('wp_statistics_dbmaint_hook', 'daily', DateTime::get('tomorrow midnight', 'U'));
         }
     }