+++ src/Service/Admin/NoticeHandler/GeneralNotices.php
@@ -125,11 +125,7 @@ class GeneralNotices
      */
     private function checkTrackingMode()
     {
-        if (Notice::isNoticeDismissed('deprecate_server_side_tracking')) {
-            return;
-        }
-
-        if (!Menus::in_plugin_page()) {
+        if (Notice::isNoticeDismissed('deprecate_server_side_tracking_14.17')) {
             return;
         }
 
@@ -139,20 +135,13 @@ class GeneralNotices
             return;
         }
 
-        $settingsUrl = Menus::admin_url('settings');
-        $noticeText  = sprintf(
-            wp_kses(
-            /* translators: %s: settings URL */
-                __('<b>WP Statistics Notice:</b> Server Side Tracking is less accurate and will be deprecated in <b>version 15</b>. Please switch to Client Side Tracking for better accuracy. <a href="%s">Update Tracking Settings</a>.', 'wp-statistics'),
-                [
-                    'b' => [],
-                    'a' => ['href' => []],
-                ]
-            ),
-            esc_url($settingsUrl)
+        $noticeText = sprintf(
+            __('<b>WP Statistics:</b> Server-Side tracking is deprecated. Please switch to Client-Side for better accuracy. <br> <a href="%1$s">Go to Tracking Settings</a> · <a href="%2$s" target="_blank">Read the Deprecation Guide</a>', 'wp-statistics'),
+            esc_url(Menus::admin_url('settings', ['row' => 'tracking_method_tr'])),
+            esc_url('https://wp-statistics.com/resources/deprecating-server-side-tracking/?utm_source=wp-statistics&utm_medium=link&utm_campaign=settings')
         );
 
-        Notice::addNotice($noticeText, 'deprecate_server_side_tracking', 'warning');
+        Notice::addNotice($noticeText, 'deprecate_server_side_tracking_14.17', 'warning');
     }
 
     /**